Move pkg/drain content to pkg/k8sutil
As 'pkg/drain' is nothing else but yet another wrapper over Kubernetes
tooling.
